break out file structure
