+++ src/app/content/actions.ts
@@ -1,0 +1,4 @@
+import {createStandardAction} from 'typesafe-actions';
+
+export const openToc = createStandardAction('Content/openToc')<void>();
+export const closeToc = createStandardAction('Content/closeToc')<void>();
