+++ src/app/content/components/Content.tsx
@@ -1,0 +1,1 @@
+export default () => null;
