+++ src/app/content/components/index.ts
@@ -1,0 +1,5 @@
+import Content from './Content';
+
+export {
+  Content,
+}
