+++ src/app/content/index.ts
@@ -1,0 +1,12 @@
+import * as actions from './actions';
+import * as routes from './routes';
+import * as types from './types';
+import reducer from './reducer';
+
+export {
+  types,
+  routes,
+  reducer,
+  actions,
+};
+
