+++ src/app/content/types.ts
@@ -1,0 +1,11 @@
+
+
+export interface Params {
+  bookId: string;
+  pageId: string;
+}
+
+export interface State {
+  tocOpen: boolean;
+  params?: Params,
+};
