+++ src/app/errors/components/index.ts
@@ -1,0 +1,5 @@
+import PageNotFound from './PageNotFound';
+
+export {
+  PageNotFound,
+}
