+++ src/app/errors/index.ts
@@ -1,0 +1,13 @@
+import * as types from './types';
+import * as routes from './routes';
+import reducer from './reducer';
+
+const actions = {};
+
+export {
+  actions,
+  routes,
+  types,
+  reducer
+}
+
