+++ src/app/errors/types.ts
@@ -1,0 +1,3 @@
+export interface State {
+  code?: number
+};
