+++ src/app/navigation/components/index.ts
@@ -1,0 +1,5 @@
+import NavigationProvider from './NavigationProvider';
+
+export {
+  NavigationProvider,
+}
