+++ src/app/selectors.ts
@@ -1,3 +1,3 @@
-import {AppState} from './app';
+import {AppState} from '.';
 
 export const localState = (state: RootState): AppState => state;
